backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
flask-cors==6.0.2
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.44
Werkzeug==3.1.4
